requirements.txt
Django
requests
django-extensions
djangorestframework-simplejwt
markdown
python-dotenv
django-filter
coverage